lab7-mattermost-bot-for-agents/pyproject.toml
[project]
name = "mattermost-a2a-bot"
version = "0.1.0"
description = "Mattermost bot для работы с A2A агентом"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mattermostdriver>=7.0.0",
    "a2a",
    "httpx>=0.25.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "openinference-instrumentation>=0.1.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []

